New patcher file for NIME and some bug fixes
For some reason the order of the arguments to JsonToCsd was changed and
also code changed so that it didn’t work with “run”, just export.
Changed code back. I find it weird not to have “mode” as the first
argument instead of the last, but there might be a good reason for it

Reverb got new min. value and PitchShifter new default value for mix

MidiCCExample.csd is not needed anymore, so this was deleted from the
examples.

NIME2017-Patch_V2.json is a suggestion for a new demo patch for the
paper, though I prefer the first one as it is more engaging to play
with.